SCP-4636 Renamed `marlowe-scaling` to `marlowe-app`.
