+++ marlowe-apps/app/Main.hs
@@ -33,9 +33,9 @@ import Language.Marlowe.Core.V1.Semantics.Types
   , Value(ChoiceValue)
   )
 import Language.Marlowe.Core.V1.Semantics.Types.Address (deserialiseAddress)
+import Language.Marlowe.Runtime.App (handle)
+import Language.Marlowe.Runtime.App.Types (Config, MarloweRequest(..), MarloweResponse(..))
 import Language.Marlowe.Runtime.ChainSync.Api (Address(unAddress), fromBech32)
-import Language.Marlowe.Runtime.Client (handle)
-import Language.Marlowe.Runtime.Client.Types (Config, MarloweRequest(..), MarloweResponse(..))
 import Language.Marlowe.Runtime.Core.Api (ContractId, MarloweVersionTag(V1))
 import Observe.Event (EventBackend, addField, hoistEvent, hoistEventBackend, withEvent, withSubEvent)
 import Observe.Event.Dynamic (DynamicEvent, DynamicEventSelector(..))
